# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Call Attribution Platform API",
    "description": "Deterministic call attribution: number allocation, ingestion, qualification, publication and administration. See specs/001-call-attribution-platform/contracts/ for the narrative wire contracts this schema implements.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/admin/audit": {
      "get": {
        "tags": [
          "AdminAudit"
        ],
        "parameters": [
          {
            "name": "target_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/admin/numbers/{id}/suspend": {
      "post": {
        "tags": [
          "AdminNumbers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/numbers/{id}/retire": {
      "post": {
        "tags": [
          "AdminNumbers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/numbers/{id}/reactivate": {
      "post": {
        "tags": [
          "AdminNumbers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/numbers/{id}/move": {
      "post": {
        "tags": [
          "AdminNumbers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveNumberRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveNumberRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveNumberRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MoveNumberRequestDto": {
        "type": "object",
        "properties": {
          "target_pool_id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}